Changed globe.pointCoordinate3D to no longer snap
